refactor: split subnet context file reader into a new class
