+++ src/vpcStatefulCidrBlockAssigner.ts
@@ -1,9 +1,8 @@
-import * as fs from 'fs';
-import * as path from 'path';
 import { IAspect, aws_ec2 as ec2 } from 'aws-cdk-lib';
 import { IConstruct } from 'constructs';
 import * as errors from './errors';
 import { SubnetCidrBlockMatcher } from './subnetCidrBlockMatcher';
+import { SubnetContextFileReader } from './subnetContextFileReader';
 import { SubnetMutator } from './subnetMutator';
 
 /**
@@ -106,10 +105,11 @@ export class VpcStatefulCidrBlockAssigner implements IAspect {
 
   private vpcCount: number = 0;
   private disallowedAvailabilityZones: Array<string> = [];
+  private readonly subnetContextFileReader: SubnetContextFileReader = new SubnetContextFileReader();
 
   constructor(props: VpcStatefulCidrBlockAssignerProps) {
-    const subnetContextFilePath = this.generateContextFilePath(props.vpcId, props.contextFileDirectory);
-    this.subnetContext = this.readSubnetContextFromFile(subnetContextFilePath);
+    const subnetContextFilePath = this.subnetContextFileReader.generateContextFilePath(props.vpcId, props.contextFileDirectory);
+    this.subnetContext = this.subnetContextFileReader.readSubnetContextFromFile(subnetContextFilePath);
     this.assignedCiderBlock = this.collectAssignedCidrBlocks(this.subnetContext);
 
     if (typeof props.availabilityZoneSubstitutions !== 'undefined') {
@@ -148,39 +148,6 @@ export class VpcStatefulCidrBlockAssigner implements IAspect {
     return newSubnetContext;
   }
 
-  private generateContextFilePath(vpcId: string, contextFileDirectory?: string): string {
-    const directoryPath: string =
-      typeof contextFileDirectory !== 'undefined' ? contextFileDirectory : process.cwd();
-    const contextFilePath = path.join(directoryPath, `${vpcId}.subnets.context.json`);
-    return contextFilePath;
-  }
-
-  private readSubnetContextFromFile(subnetContextFilePath: string): Array<SubnetRecord> {
-    let subnetContextJsonString: string;
-    let subnetContext: Array<SubnetRecord>;
-
-    if (!fs.existsSync(subnetContextFilePath)) {
-      throw errors.SUBNET_CONTEXT_FILE_DOES_NOT_EXIST;
-    }
-
-    try {
-      subnetContextJsonString = fs.readFileSync(subnetContextFilePath, 'utf-8');
-    } catch (error) {
-      throw errors.READING_SUBNET_CONTEXT_FILE;
-    }
-
-    if (subnetContextJsonString.length == 0) {
-      throw errors.EMPTY_SUBNET_CONTEXT_FILE;
-    }
-
-    try {
-      subnetContext = JSON.parse(subnetContextJsonString) as Array<SubnetRecord>;
-    } catch (error) {
-      throw errors.PARSING_SUBNET_CONTEXT_FILE;
-    }
-
-    return subnetContext;
-  }
 
   visit(node: IConstruct): void {
     if (node instanceof ec2.Vpc) {
